reworks event manager class
simplifies how the event manager works and is used.
when triggering an event, the user can now pass in n amount of arguments,
the first one being the name of the event to trigger.
this is more in line with how other (JQuery) event handlers works as well
as being more intuetive.