adds additional functionality to the world class
adds functionality which allows the user to get all entities in a world
unfiltered, by passing null as the first parameter to getEntities.

adds two new methods; getEntitiesExact and getEntitiesExcept.
the first gets the entities exactly consisting of the components passed in,
ignoring and entity that has any component other than those specified.
the second method gets all entities that does not have the components specified