exchanges Jasmine for Mocha as testing framework
since it was not possible to test EC6/Harmony features via node using Jasmine
(since it was running on Phantomjs), we are now using Mocha for unit testing.

This also makes it able to us to use Blanketjs for code coverage in an easy way,
helping towards closing issue #3.

since node needs to be run with --harmony flag, it is not possible to run the
tests as a grunt task. instead a npm script has been created, which is invoked by
typing "npm test" which then runs the Mocha grunt task

also fixed some tests and removed exceptions from code for better perforamnce and a
more unified behaviour across the API