removes mocha grunt task for npm test script and cloeses #4
since i wasn't able to get code coverage working as a grunt task
because of es6/harmony features used in the code, the grunt task
has been dropped in favor of the npm script "npm test".

this is also in line with the idea to drop grunt in favor of npm
scripts all together as a build pipe.