adds new es6 src and test files
also adds some testing scripts and cleans up the repo a bit

the new coding style focuses on clearity and readability as oppsed
to the old code which was focused on performance only.

the reasoning behind this is as javscript engines implement the es6
spec performance will get better over time and babel 5to6 can be phased
out.