changes default value of returnDetails to be false and closes #8
the various methods that has the optional parameter returnDetails should have a
it as a default value of false instead of true.

this is in order with how the system is inteded to be used, where an entity and
a component is only an id to the outside world and handled entirely by the
entitymanager.