changes system manager to have systems as map instead of arrays
since there needs to be a way to track systems, similarly to how components
are tracked - via id's, there needs to be a more logical way of storing them
that supports this is a straight forward way.

to make the API as unified as possible it will work very much like how
components are registered, removed etc.